See description
Added randomize_original argument, as it would automatically
sample/shuffle the output, which is not useful for timeseries.
Added conversion to pd.DataFrame for pd.Series and np.ndarray.
Changed order of inputs to blend() so amount is 0 for original and 1
for new.
Made sure data is the same size as keep_data when append is True.